requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
MySQL-python==1.2.5
Werkzeug==0.9.6
pylibmc==1.4.1
django-heroku-memcacheify==0.8
django-debug-toolbar==1.2.2
django-extensions==1.4.9
twython==3.2.0
-r requirements/production.txt